requirements.txt
# Core ML/DL Libraries
torch>=1.8.0       # PyTorch - Install version compatible with your CUDA, see https://pytorch.org/
transformers>=4.0.0 # Hugging Face Transformers for models/processors
numpy>=1.19.0       # Fundamental package for numerical computing

# Image/Video Processing & Visualization
opencv-python>=4.5.0 # OpenCV for image loading, resizing, filtering
Pillow>=8.0.0        # PIL fork for image manipulation (used by some scripts)
matplotlib>=3.3.0    # For plotting visualizations (json_vis.py, mask_distribution.py)
open3d>=0.13.0       # For 3D visualization (depth_3D.py)

# Utilities
tqdm>=4.50.0         # Progress bars (used in mask_eval.py)